ログアウト時の通知バーの表示ロジックの修正
